Do not post data for curl even if the data is an empty string
If the argument list contains `--data-binary`, curl will specify the content
type of a GET request even if the data is empty:

```
Content-Length: 0
Content-Type: application/x-www-form-urlencoded
```

Some sites, e.g., `elpa.emacs-china.org`, don't like this header in the GET
request.  To solve this problem, don't put `--data-binary` in the argument list
if `url-request-data` is either `nil` or an empty string.